requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.23.1
Werkzeug==2.0.3
SQLAlchemy==1.4.41
setuptools>=58.0.4
